csDetector_New_Param/requirement.txt
annotated-types==0.5.0
asttokens==2.4.1
attrs==23.1.0
backcall==0.2.0
blis==0.7.10
catalogue==2.0.9
certifi==2023.7.22
chardet==3.0.4
clean-text==0.6.0
click==8.1.7
comm==0.2.2
confection==0.1.2
contourpy==1.1.1
convokit==2.5.3
cycler==0.11.0
cymem==2.0.7
debugpy==1.8.8
decorator==5.1.1
dill==0.3.7
emoji==1.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl
et-xmlfile==1.1.0
executing==2.1.0
Flask==2.1.2
fonttools==4.54.1
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.2
idna==2.10
imbalanced-learn==0.9.0
imblearn==0.0
importlib-metadata==6.8.0
importlib_resources==6.4.5
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.2
joblib==0.17.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.5
langcodes==3.3.0
MarkupSafe==2.1.3
matplotlib==3.7.5
matplotlib-inline==0.1.7
msgpack==1.0.5
msgpack-numpy==0.4.8
murmurhash==1.0.9
nest-asyncio==1.6.0
networkx==2.4
nltk==3.8.1
numpy==1.22.0
openpyxl==3.0.9
packaging==23.1
pandas==1.4.0
parso==0.8.4
pathy==0.10.2
pause==0.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.3.0
preshed==3.0.8
progress==1.5
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.18.0
pyparsing==3.1.1
pytest==7.1.2
python-dateutil==2.9.0.post0
pytz==2023.3.post1
PyYAML==5.3.1
pyzmq==26.2.0
regex==2023.8.8
requests==2.23.0
scikit-learn==1.0.2
scipy==1.10.1
sentistrength==0.0.8
six==1.16.0
smart-open==6.4.0
smmap==5.0.0
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
stack-data==0.6.3
strsimpy==0.1.9
thinc==8.1.12
threadpoolctl==3.2.0
tomli==2.0.1
tornado==6.4.1
tqdm==4.66.1
traitlets==5.14.3
typer==0.9.0
typing_extensions==4.7.1
Unidecode==1.3.6
urllib3==1.25.11
wasabi==1.1.2
wcwidth==0.2.6
Werkzeug==2.3.7
xlrd==2.0.1
zipp==3.16.2
